+++ src/main/java/blackjack/domain/Results.java
@@ -1,9 +1,10 @@
-package blackjack.domain.results;
+package blackjack.domain;
 
 import blackjack.domain.participant.Dealer;
 import blackjack.domain.participant.Participant;
 import blackjack.domain.participant.Player;
 import blackjack.domain.participant.Players;
+import blackjack.dto.ResultsDto;
 
 import java.util.stream.Collectors;
