+++ src/main/java/blackjack/domain/exception/IllegalNameException.java
@@ -1,4 +1,4 @@
-package blackjack.utils;
+package blackjack.domain.exception;
 
 public class IllegalNameException extends RuntimeException {
     public IllegalNameException(String message) {
