+++ src/test/java/blackjack/domain/participant/PlayersTest.java
@@ -1,7 +1,7 @@
 package blackjack.domain.participant;
 
 import blackjack.utils.FixedCardDeck;
-import blackjack.utils.IllegalNameException;
+import blackjack.domain.exception.IllegalNameException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
